+++ fixtures/media-library/CHANGELOG.md
@@ -1,5 +1,21 @@
 # media-library-fixture
 
+## 1.0.32
+
+### Patch Changes
+
+- The following workspace dependencies were updated
+  - dependencies
+    - @sanity/cli bumped to 8.9.0
+
+## 1.0.31
+
+### Patch Changes
+
+- The following workspace dependencies were updated
+  - dependencies
+    - @sanity/cli bumped to 8.8.0
+
 ## 1.0.30
 
 ### Patch Changes